feat: add entity collection and parsing for various file types
- Implemented `collectEntities` for PHP, SQL, TypeScript, and JavaScript to gather relevant entities in a single pass.
- Added `parseCss` to extract CSS rules, at-rules, and custom properties while handling regions.
- Introduced `parseKdl` for parsing KDL documents, supporting node extraction and nested structures.
- Created `parseMarkdown` to identify and structure Markdown headings with configurable styles.
- Enhanced syntax highlighting management in `syntaxHighlighting.ts` to respect user settings for enabling/disabling features.
