+++ package.json
@@ -2,7 +2,7 @@
 	"name": "theToyBox",
 	"displayName": "The Toy Box",
 	"description": "The Toy Box is a VS Code extension that bundles essential productivity tools: auto-cleanup on save, a smarter code outline for SQL/PHP/Web, custom comment highlighting, GitHub-style Markdown alerts, JSON formatting, equals alignment, and real-time HTML/XML tag renaming — all in one package.",
-	"version": "0.0.20",
+	"version": "0.0.21",
 	"publisher": "ThomasDromgoole",
 	"repository": {
 		"type": "git",
@@ -410,6 +410,11 @@
 			{
 				"title": "Syntax Highlighting",
 				"properties": {
+					"theToyBox.syntaxHighlighting.enabled": {
+						"type": "boolean",
+						"default": true,
+						"description": "Master toggle for decoration-based syntax highlighting. When disabled, no decoration types are created and all per-language settings are ignored."
+					},
 					"theToyBox.syntaxHighlighting.asp": {
 						"type": "boolean",
 						"default": true,
