+++ src/outline.ts
@@ -1,57 +1,8 @@
 ﻿import * as vscode from "vscode";
-
-/**
- * Scan forward from startLine counting braces to find the closing `}` of a CSS block.
- */
-function findCssBlockEnd(
-	document: vscode.TextDocument,
-	startLine: number,
-): number {
-	let depth = 0;
-	let started = false;
-	for (let i = startLine; i < document.lineCount; i++) {
-		const text = document.lineAt(i).text;
-		for (const ch of text) {
-			if (ch === "{") {
-				depth++;
-				started = true;
-			} else if (ch === "}" && started) {
-				depth--;
-				if (depth === 0) {
-					return i;
-				}
-			}
-		}
-	}
-	return document.lineCount - 1;
-}
-
-/**
- * Starting from the line where a jQuery handler is detected, scan forward
- * counting braces to find the line of the closing `});`.
- */
-function findJqueryHandlerEnd(
-	document: vscode.TextDocument,
-	startLine: number,
-): number {
-	let depth = 0;
-	let started = false;
-	for (let i = startLine; i < document.lineCount; i++) {
-		const text = document.lineAt(i).text;
-		for (const ch of text) {
-			if (ch === "{") {
-				depth++;
-				started = true;
-			} else if (ch === "}" && started) {
-				depth--;
-				if (depth === 0) {
-					return i;
-				}
-			}
-		}
-	}
-	return document.lineCount - 1;
-}
+import { parseMarkdown } from "./outline/parseMarkdown";
+import { parseCss } from "./outline/parseCss";
+import { parseKdl } from "./outline/parseKdl";
+import { collectEntities } from "./outline/collectEntities";
 
 export class BetterOutlineProvider implements vscode.WebviewViewProvider {
 	public static readonly viewType = "betterOutlineView";
@@ -257,33 +208,11 @@ export class BetterOutlineProvider implements vscode.WebviewViewProvider {
 		document: vscode.TextDocument,
 		symbols: vscode.DocumentSymbol[],
 	) {
-		// 1. Setup & Configuration
-		const config = vscode.workspace.getConfiguration(
-			"theToyBox.customComments",
-		);
-		const stylingEnabled = config.get<boolean>("enabled", true);
-		const commentColors =
-			config.get<{ [key: string]: string }>("colors") || {};
-		const labels = config.get<{ [key: string]: string }>("labels") || {};
-		const showBackground = config.get<boolean>("showBackground", true);
-
 		const rootItems: any[] = [];
 		const regionStack: any[] = [];
 		const claimedLines = new Set<number>();
-		const triggerChars = Object.keys(commentColors);
-		const commentPrefixes = ["//", "--", "#", "%", "'"];
-
-		const allComments: any[] = [];
-		const sqlEntities: any[] = [];
-		const phpFunctions: any[] = [];
-		const tsJsItems: any[] = [];
-		const cssItems: any[] = [];
 
 		const lang = document.languageId.toLowerCase();
-		const isSqlFile = ["sql", "postgresql", "mssql", "postgres"].includes(
-			lang,
-		);
-		const isPhpFile = lang === "php";
 		const isMarkdownFile = lang === "markdown";
 		const isCssFile = ["css", "scss", "less"].includes(lang);
 		const isTsJsFile = [
@@ -297,663 +226,22 @@ export class BetterOutlineProvider implements vscode.WebviewViewProvider {
 
 		// 2a. Markdown: parse headings into a nested tree
 		if (isMarkdownFile) {
-			const headingConfig = vscode.workspace.getConfiguration(
-				"theToyBox.markdownHeadings",
-			);
-			const headingEnabled = headingConfig.get<boolean>("enabled", true);
-			const headingColors =
-				headingConfig.get<{ [key: string]: string }>("colors") || {};
-			const headingShowBackground = headingConfig.get<boolean>(
-				"showBackground",
-				true,
-			);
-
-			const rootHeadings: any[] = [];
-			const stack: { level: number; item: any }[] = [];
-
-			for (let i = 0; i < document.lineCount; i++) {
-				const text = document.lineAt(i).text;
-				const match = text.match(/^(#{1,6})\s+(.*)/);
-				if (!match) {
-					continue;
-				}
-
-				const level = match[1].length;
-				const levelKey = `h${level}`;
-				const color = headingEnabled ? headingColors[levelKey] : null;
-
-				const item: any = {
-					label: match[2].trim(),
-					line: i,
-					isMarkdownHeading: true,
-					headingLevel: level,
-					color: color || null,
-					backgroundColor:
-						color && headingShowBackground ? color + "33" : null,
-					isBold: level === 1,
-					isRegion: false,
-					children: [],
-				};
-
-				// Pop the stack until we find a heading of a higher level
-				while (
-					stack.length > 0 &&
-					stack[stack.length - 1].level >= level
-				) {
-					stack.pop();
-				}
-
-				if (stack.length > 0) {
-					const parent = stack[stack.length - 1].item;
-					parent.children.push(item);
-				} else {
-					rootHeadings.push(item);
-				}
-
-				stack.push({ level, item });
-			}
-
-			return rootHeadings;
+			return parseMarkdown(document);
 		}
 
 		// 2a-css. CSS/SCSS/Less: parse at-rules, selectors, custom properties, #regions
 		if (isCssFile) {
-			const atRuleRe =
-				/^\s*(@(?:media|keyframes|mixin|include|layer|supports|font-face|charset|import|use|forward|each|for|if|else|while|function|return|debug|warn|error))\b([^{;]*)/i;
-			const selectorRe =
-				/^\s*([.#]?[\w][\w\s.#>+~[\]:()=*^$|"'-]*)(?:\s*,\s*[\w.#][\w\s.#>+~[\]:()=*^$|"'-]*)?\s*\{/;
-			const customPropRe = /^\s*(--[\w-]+)\s*:/;
-			const regionStartRe = /^[#\/\-\*\s!]*region\s+(.*)/i;
-			const regionEndRe = /^[#\/\-\*\s!]*endregion/i;
-
-			const cssRegionStack: any[] = [];
-			const cssRootItems: any[] = [];
-
-			const pushCssItem = (item: any) => {
-				if (cssRegionStack.length > 0) {
-					cssRegionStack[cssRegionStack.length - 1].children.push(
-						item,
-					);
-				} else {
-					cssRootItems.push(item);
-				}
-			};
-
-			for (let i = 0; i < document.lineCount; i++) {
-				const text = document.lineAt(i).text;
-				const trimmed = text.trim();
-
-				// Check for region markers before skipping comment lines
-				const regionStartMatch = trimmed.match(regionStartRe);
-				if (regionStartMatch) {
-					const label =
-						regionStartMatch[1].replace(/\*\/\s*$/, "").trim() ||
-						"Region";
-					cssRegionStack.push({
-						label,
-						line: i,
-						isRegion: true,
-						children: [],
-					});
-					continue;
-				}
-				if (regionEndRe.test(trimmed)) {
-					if (cssRegionStack.length > 0) {
-						const region = cssRegionStack.pop();
-						if (cssRegionStack.length > 0) {
-							cssRegionStack[
-								cssRegionStack.length - 1
-							].children.push(region);
-						} else {
-							cssRootItems.push(region);
-						}
-					}
-					continue;
-				}
-
-				if (
-					!trimmed ||
-					trimmed.startsWith("//") ||
-					trimmed.startsWith("/*") ||
-					trimmed.startsWith("*")
-				) {
-					continue;
-				}
-
-				const atMatch = trimmed.match(atRuleRe);
-				if (atMatch) {
-					const keyword = atMatch[1].toLowerCase();
-					const detail = atMatch[2].trim();
-					const label = detail ? `${keyword} ${detail}` : keyword;
-					const endLine = trimmed.endsWith("{")
-						? findCssBlockEnd(document, i)
-						: i;
-					pushCssItem({
-						label,
-						line: i,
-						endLine,
-						cssType: keyword.slice(1), // strip @
-						isRegion: endLine > i,
-						children: [],
-					});
-					continue;
-				}
-
-				const propMatch = trimmed.match(customPropRe);
-				if (propMatch) {
-					pushCssItem({
-						label: propMatch[1],
-						line: i,
-						cssType: "customProp",
-						isRegion: false,
-						children: [],
-					});
-					continue;
-				}
-
-				if (selectorRe.test(trimmed)) {
-					// Collapse multi-line selectors to first 60 chars
-					const label = trimmed
-						.replace(/\s*\{.*$/, "")
-						.trim()
-						.slice(0, 60);
-					const selectorKind = trimmed.trimStart().startsWith("#")
-						? "id"
-						: trimmed.trimStart().startsWith(".")
-							? "class"
-							: "element";
-					pushCssItem({
-						label,
-						line: i,
-						endLine: findCssBlockEnd(document, i),
-						cssType: "selector",
-						cssSelector: selectorKind,
-						isRegion: false,
-						children: [],
-					});
-				}
-			}
-
-			// Flush any unclosed regions to root
-			while (cssRegionStack.length > 0) {
-				cssRootItems.push(cssRegionStack.pop());
-			}
-
-			return cssRootItems.sort((a, b) => a.line - b.line);
+			return parseCss(document);
 		}
 
 		// 2a-kdl. KDL: parse nodes into a hierarchical tree
 		if (isKdlFile) {
-			const kdlRoot: any[] = [];
-			const kdlStack: any[] = []; // stack of open container nodes
-
-			for (let i = 0; i < document.lineCount; i++) {
-				const trimmed = document.lineAt(i).text.trim();
-
-				// Skip blank lines, line comments, and block comment lines
-				if (
-					!trimmed ||
-					trimmed.startsWith("//") ||
-					trimmed.startsWith("/*") ||
-					trimmed.startsWith("*")
-				) {
-					continue;
-				}
-
-				// Closing brace → pop the stack
-				if (trimmed === "}" || trimmed === "};") {
-					if (kdlStack.length > 0) {
-						kdlStack.pop();
-					}
-					continue;
-				}
-
-				// Extract node name: bare identifier or quoted string
-				let nodeName: string | null = null;
-				let nameEnd = 0;
-				const bareMatch = trimmed.match(/^([-\w$.:]+)/);
-				const quotedMatch = trimmed.match(/^"((?:[^"\\]|\\.)*)"/);
-				if (bareMatch) {
-					nodeName = bareMatch[1];
-					nameEnd = bareMatch[0].length;
-				} else if (quotedMatch) {
-					nodeName = `"${quotedMatch[1]}"`;
-					nameEnd = quotedMatch[0].length;
-				}
-				if (!nodeName) {
-					continue;
-				}
-
-				// Extract the first argument after the node name.
-				// - key="value" or key=bare  → "key value" (no = or quotes)
-				// - "positional string"       → the unquoted string
-				// - bare positional value     → the value as-is
-				// Stops at unquoted `{` or `//`.
-				let firstArg = "";
-				{
-					let j = nameEnd;
-					// skip whitespace
-					while (
-						j < trimmed.length &&
-						(trimmed[j] === " " || trimmed[j] === "\t")
-					) {
-						j++;
-					}
-					// stop immediately at block opener or comment
-					if (
-						j < trimmed.length &&
-						trimmed[j] !== "{" &&
-						!(trimmed[j] === "/" && trimmed[j + 1] === "/")
-					) {
-						// skip type annotation e.g. (u8)
-						if (trimmed[j] === "(") {
-							while (j < trimmed.length && trimmed[j] !== ")") {
-								j++;
-							}
-							j++; // skip ')'
-							while (
-								j < trimmed.length &&
-								(trimmed[j] === " " || trimmed[j] === "\t")
-							) {
-								j++;
-							}
-						}
-						const identMatch = trimmed
-							.slice(j)
-							.match(/^([-\w$.:]+)/);
-						if (identMatch) {
-							const ident = identMatch[1];
-							j += ident.length;
-							if (trimmed[j] === "=") {
-								j++; // skip '='
-								let val = "";
-								if (trimmed[j] === '"') {
-									// quoted value — strip the quotes
-									j++;
-									while (
-										j < trimmed.length &&
-										trimmed[j] !== '"'
-									) {
-										if (trimmed[j] === "\\") {
-											j++;
-										}
-										if (j < trimmed.length) {
-											val += trimmed[j++];
-										}
-									}
-									// skip closing quote
-								} else {
-									// bare value (number, bool, identifier)
-									const valMatch = trimmed
-										.slice(j)
-										.match(/^[^\s{;\/]+/);
-									if (valMatch) {
-										val = valMatch[0];
-									}
-								}
-								firstArg = val ? `${ident} ${val}` : ident;
-							} else {
-								// positional bare value
-								firstArg = ident;
-							}
-						} else if (trimmed[j] === '"') {
-							// positional quoted string — strip the quotes
-							j++;
-							let val = "";
-							while (j < trimmed.length && trimmed[j] !== '"') {
-								if (trimmed[j] === "\\") {
-									j++;
-								}
-								if (j < trimmed.length) {
-									val += trimmed[j++];
-								}
-							}
-							firstArg = val;
-						}
-					}
-				}
-
-				const label = firstArg ? `${nodeName} ${firstArg}` : nodeName;
-
-				// Detect whether this node opens a multi-line children block.
-				// A line with `{` but no matching `}` opens a persistent scope.
-				const hasBrace = trimmed.includes("{");
-				const hasClose = trimmed.includes("}");
-				const opensBlock = hasBrace && !hasClose;
-
-				const item: any = {
-					label,
-					line: i,
-					kdlType: "node",
-					isRegion: opensBlock,
-					children: [],
-				};
-
-				if (kdlStack.length > 0) {
-					kdlStack[kdlStack.length - 1].children.push(item);
-				} else {
-					kdlRoot.push(item);
-				}
-
-				if (opensBlock) {
-					kdlStack.push(item);
-				}
-			}
-
-			return kdlRoot;
+			return parseKdl(document);
 		}
 
-		// 2. Pass 1: Identify Comments & SQL Entities (Tables, Procs, Funcs, Views) & PHP Functions
-		for (let i = 0; i < document.lineCount; i++) {
-			const line = document.lineAt(i);
-			const text = line.text;
-			const trimmed = text.trim();
-
-			// --- PHP Function Detection ---
-			if (isPhpFile) {
-				// Match PHP function declarations: function functionName(...) {
-				const phpFuncMatch = text.match(
-					/^\s*(public|private|protected|static|async)?\s*function\s+(\w+)\s*\(/,
-				);
-
-				if (phpFuncMatch) {
-					phpFunctions.push({
-						label: phpFuncMatch[2] + "()",
-						line: i,
-						phpType: "function",
-						isRegion: false,
-						children: [],
-					});
-					continue;
-				}
-			}
-
-			// --- TypeScript / JavaScript Detection ---
-			if (isTsJsFile) {
-				// Classes: TS only — JS outline is limited to functions, jQuery events, and comments
-				if (!isJsFile) {
-					const classMatch = text.match(
-						/^\s*(?:export\s+(?:default\s+)?)?(?:abstract\s+)?class\s+(\w+)/,
-					);
-					if (classMatch) {
-						tsJsItems.push({
-							label: classMatch[1],
-							line: i,
-							tsJsType: "class",
-							isRegion: false,
-							children: [],
-						});
-						continue;
-					}
-				}
-
-				const funcMatch = text.match(
-					/^\s*(?:export\s+(?:default\s+)?)?(?:async\s+)?function\s+(\w+)\s*\(/,
-				);
-				if (funcMatch) {
-					tsJsItems.push({
-						label: funcMatch[1] + "()",
-						line: i,
-						tsJsType: "function",
-						isRegion: false,
-						children: [],
-					});
-					continue;
-				}
-
-				const arrowMatch = text.match(
-					/^\s*(?:export\s+)?(?:const|let|var)\s+(\w+)\s*=\s*(?:async\s+)?(?:\([^)]*\)|\w+)\s*=>/,
-				);
-				if (arrowMatch) {
-					tsJsItems.push({
-						label: arrowMatch[1] + "()",
-						line: i,
-						tsJsType: "arrowFunction",
-						isRegion: false,
-						children: [],
-					});
-					continue;
-				}
-
-				// jQuery .on() — $("#id").on("event", fn) or $(".cls").on("event", fn)
-				const jqOnMatch = text.match(
-					/\$\(["']([.#][\w-]+)["']\)\.on\(["'](\w+)["']\s*,\s*function/,
-				);
-
-				if (jqOnMatch) {
-					const raw = jqOnMatch[1];
-					const isId = raw.startsWith("#");
-					const name = raw.slice(1);
-					const event =
-						jqOnMatch[2].charAt(0).toUpperCase() +
-						jqOnMatch[2].slice(1);
-					tsJsItems.push({
-						label: `${name}.${event}`,
-						line: i,
-						endLine: findJqueryHandlerEnd(document, i),
-						tsJsType: "jqueryEvent",
-						jquerySelector: isId ? "id" : "class",
-						isRegion: false,
-						children: [],
-					});
-					continue;
-				}
-
-				// jQuery .delegate() — $(...).delegate(".cls", "event", fn) or $(...).delegate("#id", "event", fn)
-				const jqDelegateMatch = text.match(
-					/\.delegate\(["']([.#][\w-]+)["']\s*,\s*["'](\w+)["']\s*,\s*function/,
-				);
-
-				if (jqDelegateMatch) {
-					const raw = jqDelegateMatch[1];
-					const isId = raw.startsWith("#");
-					const name = raw.slice(1);
-					const event =
-						jqDelegateMatch[2].charAt(0).toUpperCase() +
-						jqDelegateMatch[2].slice(1);
-					tsJsItems.push({
-						label: `${name}.${event}.Delegate`,
-						line: i,
-						endLine: findJqueryHandlerEnd(document, i),
-						tsJsType: "jqueryEvent",
-						jquerySelector: isId ? "id" : "class",
-						isRegion: false,
-						children: [],
-					});
-					continue;
-				}
-			}
-
-			// --- SQL Entity Detection ---
-			if (isSqlFile) {
-				// Regex to match "CREATE [OR REPLACE] TABLE/VIEW/PROCEDURE/FUNCTION name"
-				const tableMatch = text.match(/create\s+table\s+([\w\.]+)/i);
-				const viewMatch = text.match(
-					/create\s+(?:or\s+replace\s+)?view\s+([\w\.]+)/i,
-				);
-				const procMatch = text.match(
-					/create\s+(?:or\s+replace\s+)?(?:stored\s+)?procedure\s+([\w\.]+)/i,
-				);
-				const funcMatch = text.match(
-					/create\s+(?:or\s+replace\s+)?function\s+([\w\.]+)/i,
-				);
-
-				if (tableMatch) {
-					sqlEntities.push({
-						label: tableMatch[1],
-						line: i,
-						sqlType: "table",
-						isRegion: false,
-						children: [],
-					});
-					continue;
-				} else if (procMatch) {
-					sqlEntities.push({
-						label: procMatch[1],
-						line: i,
-						sqlType: "procedure",
-						isRegion: false,
-						children: [],
-					});
-					continue;
-				} else if (funcMatch) {
-					sqlEntities.push({
-						label: funcMatch[1],
-						line: i,
-						sqlType: "function",
-						isRegion: false,
-						children: [],
-					});
-					continue;
-				} else if (viewMatch) {
-					sqlEntities.push({
-						label: viewMatch[1],
-						line: i,
-						sqlType: "view",
-						isRegion: false,
-						children: [],
-					});
-					continue;
-				}
-			}
-
-			// --- Comment Scanning with String-Safe Logic (O(n) flag-based) ---
-			let foundPrefix: string | undefined;
-			let prefixIndex = -1;
-
-			let inSingle = false;
-			let inDouble = false;
-			let hasNonWhitespaceBefore = false;
-			for (let charIdx = 0; charIdx < text.length; charIdx++) {
-				const ch = text[charIdx];
-
-				// Track closing of open string literals
-				if (inDouble) {
-					if (ch === "\\") {
-						charIdx++;
-						continue;
-					}
-					if (ch === '"') {
-						inDouble = false;
-					}
-					continue;
-				}
-				if (inSingle) {
-					if (ch === "\\") {
-						charIdx++;
-						continue;
-					}
-					if (ch === "'") {
-						inSingle = false;
-					}
-					continue;
-				}
-
-				// Outside any string: check for a comment prefix
-				const currentPrefix = commentPrefixes.find((p) =>
-					text.startsWith(p, charIdx),
-				);
-				if (currentPrefix) {
-					// Strict check for SQL ' and # (must start the line)
-					if (currentPrefix === "'" || currentPrefix === "#") {
-						if (hasNonWhitespaceBefore) {
-							// Rejected as comment; ' can still open a string
-							if (ch === "'") {
-								inSingle = true;
-							}
-							continue;
-						}
-						const after = text.substring(charIdx + 1);
-						if (
-							!triggerChars.some((s) =>
-								after.trimStart().startsWith(s),
-							) &&
-							!after.startsWith(" ")
-						) {
-							// Rejected as comment; ' can still open a string
-							if (ch === "'") {
-								inSingle = true;
-								hasNonWhitespaceBefore = true;
-							}
-							continue;
-						}
-					}
-					prefixIndex = charIdx;
-					foundPrefix = currentPrefix;
-					break;
-				}
-
-				// Track string openings and non-whitespace
-				if (ch === '"') {
-					inDouble = true;
-					hasNonWhitespaceBefore = true;
-				} else if (ch === "'") {
-					inSingle = true;
-					hasNonWhitespaceBefore = true;
-				} else if (ch !== " " && ch !== "\t") {
-					hasNonWhitespaceBefore = true;
-				}
-			}
-
-			if (!foundPrefix || prefixIndex === -1) {
-				continue;
-			}
-
-			// Skip region markers
-			if (
-				/^[#\/\-\*\s!]*region\s+/i.test(trimmed) ||
-				/^[#\/\-\*\s!]*endregion/i.test(trimmed)
-			) {
-				continue;
-			}
-
-			const afterPrefix = text
-				.substring(prefixIndex + foundPrefix.length)
-				.trimStart();
-			const triggerChar = triggerChars.find((char) => {
-				if (!afterPrefix.startsWith(char)) {
-					return false;
-				}
-
-				// For PHP dollar signs, don't treat it as a trigger if followed by a word character (variable)
-				if (char === "$") {
-					const charAfterTrigger = afterPrefix.charAt(char.length);
-					if (/\w/.test(charAfterTrigger)) {
-						return false; // It's a PHP variable, not a trigger
-					}
-				}
-
-				return true;
-			});
-
-			let displayLabel = afterPrefix;
-			if (triggerChar) {
-				const replacementWord = labels[triggerChar];
-				const contentOnly = afterPrefix
-					.substring(triggerChar.length)
-					.trimStart();
-				displayLabel = replacementWord
-					? `${replacementWord}: ${contentOnly}`
-					: `${triggerChar}: ${contentOnly}`;
-			}
-
-			allComments.push({
-				label: displayLabel,
-				line: i,
-				isComment: true,
-				color:
-					stylingEnabled && triggerChar
-						? commentColors[triggerChar]
-						: null,
-				backgroundColor:
-					stylingEnabled && triggerChar && showBackground
-						? commentColors[triggerChar] + "33"
-						: null,
-				isBold: stylingEnabled && triggerChar === "!",
-				isRegion: false,
-				children: [],
-			});
-		}
+		// 2. Pass 1: collect all entities and comments in a single pass
+		const { allComments, sqlEntities, phpFunctions, tsJsItems } =
+			collectEntities(document);
 
 		// 3. Helper for Nesting
 		const nestContent = (symbolList: vscode.DocumentSymbol[]): any[] => {
